% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{activate.HCAExplorer}
\alias{activate.HCAExplorer}
\title{Activate projects, samples, or files to display in the HCAExplorer Object}
\usage{
\method{activate}{HCAExplorer}(.data, what = c("projects", "samples", "files"))
}
\arguments{
\item{.data}{An HCAExplorer object}

\item{what}{character(1). Either 'projects', 'samples', or 'files'.}
}
\value{
An HCAExplorer object with medified activation.
}
\description{
The HCAExplorer can display its results in a variety of ways.
 Choose whether to display entries by project, samples, or files.
 The HCAExplorer class always defaults to projects.
}
\examples{
 ## Initiate an HCAExplorer object.
 x <- HCAExplorer()
 ## Display the object. Notice "projects" are shown by default.
 x

 ## Now activate "samples" and now notice that "samples" are displayed.
 x <- activate(x, 'samples')
 x

 ## Now activate "files" and now notice that "files" are displayed.
 x <- activate(x, 'files')
 x

 ## Now activate "projects" and now notice that the original "projects" are
 ## displayed.
 x <- activate(x, 'projects')
 x

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class.
}
