% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{activate}
\alias{activate}
\alias{activate,HCAExplorer-method}
\title{Activate projects, samples, or files to display in the HCAExplorer Object}
\usage{
\S4method{activate}{HCAExplorer}(hca, what = c("projects", "samples",
  "files"))
}
\arguments{
\item{hca}{An HCAExplorer object}

\item{what}{character(1). Either 'projects', 'samples', or 'files'.}
}
\value{
An HCAExplorer object with medified activation.
}
\description{
Choose whether to display entries by project, samples, or files.
 The HCAExplorer class always defaults to projects.
}
\examples{
 hca <- HCAExplorer()
 hca \%>\% activate('samples')
 hca \%>\% activate('files')

}
