% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{download-expression-matrices}
\alias{download-expression-matrices}
\alias{checkExpressionMatricesAvailability}
\alias{downloadExpressionMatrices}
\alias{checkExpressionMatricesAvailability,HCAExplorer-method}
\alias{downloadExpressionMatrices,HCAExplorer-method}
\title{Download Expression Matrices with HCAExplorer}
\usage{
checkExpressionMatricesAvailability(x, ...)
 downloadExpressionMatrices(x, ...)

\S4method{checkExpressionMatricesAvailability}{HCAExplorer}(
  x,
  format = c("loom", "csv", "mtx"),
  organism
)

\S4method{downloadExpressionMatrices}{HCAExplorer}(
  x,
  format = c("loom", "csv", "mtx"),
  organism,
  useBiocFileCache = TRUE
)
}
\arguments{
\item{x}{an HCAExplorer object. All project entries in the object will
attempt to be downloaded. Use subsetting to reduce the projects that are
to be downloaded.}

\item{format}{either "loom", "csv", of "mtx". The default value is "loom".
Which format the expression matrices should be downloaded as. If "loom", a
`LoomExperiment` object will be returned. If "csv", a tibble will be
returned. If "mtx", a `SingleCellExperiment` object will be returned.}

\item{organism}{character(1). Which organism file to download. Usually either
"Homo sapiens" or "Mus musculus". If not specified, all organism files will
be included in the download.}

\item{useBiocFileCache}{logical(1). Whether to save file with
`BiocFileCache`. Defaults to `TRUE`. If `TRUE`, `BiocFileCache` will be used
to save a persistent copy of the file to the user's computer. This
persistent copy will be used instead of redownloading the expression
matrices in subsequent uses. If `FALSE`, the file will be saved to a
temporary directory and will be removed when the session ends.}

\item{...}{Additional arguments.}
}
\value{
For `checkExpressionMatrixAvailability`, returns a tibble.
 For `downloadExpressionMatrices', returns a list of downloaded expression
 matrices in a format specified by the `format` argument.
}
\description{
Methods used to download expression matrices availiable through the HCA Data
 Portal. The method `checkExpressionMatricesAvaiability` specifies whether
 expression matrices can be downloaded for a project. The method
 `downloadExpressionMatrices` downloads and returns expression matrices in
 a variety of formats.
}
\examples{
hca <- HCAExplorer()
checkExpressionMatricesAvailability(hca)

hca <- hca[5]
hca <- downloadExpressionMatrices(hca, useBiocFileCache = FALSE)

}
\author{
Daniel Van Twisk
}
