% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{fields,HCAExplorer-method}
\alias{fields,HCAExplorer-method}
\title{List supported fields of an HCAExplorer object}
\usage{
\S4method{fields}{HCAExplorer}(x)
}
\arguments{
\item{x}{An HCAExplorer object.}
}
\value{
A tibble indicating fields that can be queried upon.
}
\description{
This function is used to discover possible fields that can be
 queried upon in an HCAExplorer object.
}
\examples{
## Intiate an HCAExplorer object.
x <- HCAExplorer()

## Find which fields can be queried upon.
fields(x)

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{values}} for how to obtain possible values of a field.
}
