% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{fields}
\alias{fields}
\alias{fields,ProjectBrowser-method}
\title{List supported fields of an ProjectBrowser object}
\usage{
\S4method{fields}{ProjectBrowser}(hca)
}
\value{
A tibble indicating fields that can be queried upon.
}
\description{
List supported fields of an ProjectBrowser object
}
\examples{
hca <- ProjectBrowser()
hca \%>\% fields

}
