% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{filter.HCAExplorer}
\alias{filter.HCAExplorer}
\title{Filter an HCAExplorer object using fields and values}
\usage{
\method{filter}{HCAExplorer}(.data, ..., .preserve)
}
\arguments{
\item{.data}{An HCAExplorer object to query.}

\item{...}{Any number of fields and values to queried against. The binary
operators '==' and '%in%' are allowed while only the combination operator
'&' is allowed. See examples.}

\item{.preserve}{Unused argument}
}
\value{
An HCAExplorer object with the desired query performed.
}
\description{
Given some amount of fields and values associated with them,
 modify the search performed by the HCAExplorer object. This function adds
 terms to the query that is ultimately performed.
}
\examples{
 hca <- HCAExplorer()
 hca \%>\% fields
 hca \%>\% values('organ')
 hca \%>\% values('disease')
 ## Perform query (organ is 'blood' OR 'brain') AND disease is 'normal'.
 hca \%>\% filter(organ == c('blood', 'brain') & disease == 'normal')
 ## Also
 hca \%>\% filter(organ == c('blood', 'brain'), disease == 'normal')

}
