% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getManifest}
\alias{getManifest}
\alias{getManifest,HCAExplorer-method}
\title{Obtain metadata information from an HCAExplorer object}
\usage{
\S4method{getManifest}{HCAExplorer}(hca, fileFormat)
}
\arguments{
\item{hca}{An HCAExplorer object}

\item{fileFormat}{character. A character vector of file formats of metadata
to obtain.}
}
\value{
a tibble of metadata information
}
\description{
Obtain metadata infromation from an HCAExplorer object.
 This metadata can then be passed on to download files from other services.
}
\examples{
 hca <- HCAExplorer()
 hca <- hca \%>\% select(projects = c(1))
 formats <- hca \%>\% getManifestFileFormats
 hca \%>\% getManifest(fileFormat = formats[1])

}
