% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getManifestFileFormats}
\alias{getManifestFileFormats}
\alias{getManifestFileFormats,HCAExplorer-method}
\title{Show all possible manifest file formats for current selection}
\usage{
\S4method{getManifestFileFormats}{HCAExplorer}(hca)
}
\arguments{
\item{hca}{An HCAExplorer object}
}
\value{
A character vecotr of information about possible file formats.
}
\description{
Show all possible manifest file formats for the
 current selection of projects in the HCAExplorer object. To be
 used in conjunction with 'getManifest()'.
}
\examples{
 hca <- HCAExplorer()
 hca \%>\% getManifestFileFormats

}
