% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HCAExplorer-class}
\alias{HCAExplorer-class}
\alias{.HCAExplorer}
\title{The Project Browser Class}
\arguments{
\item{per_page}{the number of results to display per request.}

\item{url}{character(1) The url of the Human Cell Atlas Backend to access.}
}
\description{
A class that displays Human Cell Atlas(HCA) information obtained
 from a digested version of the Human Cell Atlas referred to as the Azul
 Backend. This class allows one to view, filter, and select entries in the
 to ultimately download metadata files and then other data such as matrices.
}
\author{
Daniel Van Twisk
}
