% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azul_api.R
\docType{methods}
\name{nextResults}
\alias{nextResults}
\alias{nextResults,ProjectBrowser-method}
\title{Next Results}
\usage{
\S4method{nextResults}{ProjectBrowser}(result)
}
\value{
A ProjectBrowser object that displays the next results
}
\description{
Fetch the next set of bundles from a ProjectBrowser Object
}
\examples{

hca <- ProjectBrowser()
hca <- nextResults(hca)
hca

}
\author{
Daniel Van Twisk
}
