% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProjectView-class}
\alias{ProjectView-class}
\alias{.ProjectView}
\title{The ProjectView class}
\arguments{
\item{...}{Arguments to be passed to list constructor.}
}
\description{
A class that displays all metadata information from a list of
 projects in an HCAExplorer object. Extends lists.
}
