% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{resetQuery}
\alias{resetQuery}
\alias{resetQuery,HCAExplorer-method}
\title{Reset all queries performed on an object}
\usage{
\S4method{resetQuery}{HCAExplorer}(hca)
}
\arguments{
\item{hca}{An HCAExplorer object}
}
\value{
An HCAExplorer object with the changes applied to it.
}
\description{
Reset all queries performed on an HCAExplorer
 object then return the result.
}
\examples{
 hca <- HCAExplorer()
 hca <- hca \%>\% filter(organ == 'blood')
 hca <- hca \%>\% select(projects = c(1, 2))
 hca \%>\% resetQuery

}
