% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{resetSelect}
\alias{resetSelect}
\alias{resetSelect,HCAExplorer-method}
\title{Reset a select perform on an HCAExplorer object.}
\usage{
\S4method{resetSelect}{HCAExplorer}(x)
}
\arguments{
\item{x}{An HCAExplorer Object}
}
\value{
An HCAExplorer object with its selected slot returned to its default
 value.
}
\description{
This function will reset the selected slot of an HCAExplorer
 object to its default selection.
}
\examples{
 
 ## Initiate an HCAExplorer Object.
 x <- HCAExplorer()
 ## View object to decide which colums to select.
 x

 ## Decide to select columns 'projcts.projectTitle' and 'samples.organ'.
 x <- x \%>\% select('project.projectTitle', 'samples.organ')
 x

 ## Revert selec() with resetSelect()
 x <- resetSelect(x)
 x

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{select}} for how to make the initial selection.
}
