% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,HCAExplorer-method}
\title{Obtain search results from a HCAExplorer Object}
\usage{
\S4method{results}{HCAExplorer}(object)
}
\arguments{
\item{object}{An HCAExplorer object}
}
\value{
a tibble
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated.
}
\examples{
 hca <- HCAExplorer()
 hca \%>\% results

}
