% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{select.HCAExplorer}
\alias{select.HCAExplorer}
\title{Filter entries by position or name}
\usage{
\method{select}{HCAExplorer}(.data, ..., projects)
}
\arguments{
\item{.data}{An HCAObject to filter}

\item{...}{Unused argument}

\item{projects}{numeric or character. Either the positions in the output show
method of the HCAExplorer object or the entries' names.}
}
\value{
An HCAExplorer object with the applied filter.
}
\description{
This function works differently from other usages of select.
 This function allows the user to filter entries based on position or name.
 This function creates a filter targeting the entryId.
}
\examples{
 hca <- HCAExplorer()
 hca
 hca \%>\% select(projects = c(1, 2))

}
