% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,HCAExplorer,ANY,ANY,ANY-method}
\alias{[,HCAExplorer,ANY,ANY,ANY-method}
\alias{`[`,HCAExplorer-method}
\title{Subset an HCAExplorer Object by row number or project name}
\usage{
\S4method{[}{HCAExplorer,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An HCAExplorer object.}

\item{i}{Either a numeric vector indicating which rows to choose or a
character vector of project titles indicating which projects to choose.}

\item{j}{Unused argument.}

\item{...}{Unused argument.}

\item{drop}{Unused argument.}
}
\value{
An HCAExplorer object with the applied subset.
}
\description{
Allows subsetting an HCAExplorer object by row number or
 project name. This method internally acts as a filter and actually
 performs a query using "projectIds".
}
\examples{
 ## Initiate an HCAExplorer Object.
 x <- HCAExplorer()
 ## View object to decide which projects to choose.
 x
 ## Decide to select projects 1, 2, 3, and 7
 x <- x[c(1:3, 7)]
 x

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{filter}} for how to performm a query on an HCAExplorer object.
}
