% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{undoQuery}
\alias{undoQuery}
\alias{undoQuery,HCAExplorer-method}
\title{Undo one or multiple queries performed on an object}
\usage{
\S4method{undoQuery}{HCAExplorer}(hca, n = 1L)
}
\arguments{
\item{hca}{An HCAExplorer object}

\item{n}{integer(1). The number of queries to step back from.}
}
\value{
An HCAExplorer object with the changes applied to it.
}
\description{
Undo one or multiple queries performed on an HCAExplorer
 object then return the result.
}
\examples{
 hca <- HCAExplorer()
 hca <- hca \%>\% filter(organ == 'blood')
 hca <- hca \%>\% select(projects = c(1, 2))
 hca \%>\% undoQuery(n = 1L)

}
