% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{values,HCAExplorer-method}
\alias{values,HCAExplorer-method}
\title{List all values for certain fields in a HCAExplorer Object}
\usage{
\S4method{values}{HCAExplorer}(x, fields)
}
\arguments{
\item{x}{A HCAExplorer Object.}

\item{fields}{a character vector of fields to display avaiable values for.}

\item{...}{Other arguments.}
}
\value{
a list of possible values for a filter
}
\description{
List all values for certain fields in a HCAExplorer Object
}
\examples{
hca <- HCAExplorer()
hca \%>\% values('organ')

}
