% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{viewProjects}
\alias{viewProjects}
\alias{viewProjects,HCAExplorer-method}
\title{View all metadata about a selection of projects}
\usage{
\S4method{viewProjects}{HCAExplorer}(hca)
}
\arguments{
\item{hca}{An HCAExplorer object}
}
\value{
A list of all metadata in the selected entries
}
\description{
Returns a list of all the metadata from the current selection
 of entries in the HCAExplorer object.
}
\examples{
 hca <- HCAExplorer()
 view <- hca \%>\% viewProjects()
 view

}
