% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{activate}
\alias{activate}
\alias{activate,ProjectBrowser-method}
\title{Activate projects, samples, or files to display in the ProjectBrowser Object}
\usage{
\S4method{activate}{ProjectBrowser}(hca, what = c("projects", "samples",
  "files"))
}
\description{
Activate projects, samples, or files to display in the ProjectBrowser Object
}
