% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{fields}
\alias{fields}
\title{List supported fields of an HCAExplorer object}
\usage{
fields(x)
}
\arguments{
\item{x}{An HCAExplorer object.}
}
\value{
A tibble indicating fields that can be queried upon.
}
\description{
This function is used to discover possible fields that can be
 queried upon in an HCAExplorer object.
}
\examples{
## Intiate an HCAExplorer object.
x <- HCAExplorer()

## Find which firelds can be queried upon.
fields(x)

}
\seealso{
[HCAExplorer()] [values()]
}
