% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\docType{methods}
\name{fields}
\alias{fields}
\alias{fields,HCAExplorer-method}
\title{List supported fields of an HCAExplorer object}
\usage{
\S4method{fields}{HCAExplorer}(hca)
}
\arguments{
\item{hca}{An HCAExplorer object}
}
\value{
A tibble indicating fields that can be queried upon.
}
\description{
List supported fields of an HCAExplorer object
}
\examples{
hca <- HCAExplorer()
hca \%>\% fields

}
