% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getManifest}
\alias{getManifest}
\alias{getManifest,HCAExplorer-method}
\title{Obtain metadata information from an HCAExplorer object}
\usage{
\S4method{getManifest}{HCAExplorer}(x, fileFormat)
}
\arguments{
\item{x}{An HCAExplorer object}

\item{fileFormat}{character. A character vector of file formats of metadata
to obtain. The possible aruments can be found using the
getManifestFileFormats method.}
}
\value{
a tibble of metadata information.
}
\description{
Obtain metadata infromation from an HCAExplorer object.
 This metadata can then be passed on to download files from other services.
}
\examples{
 ## Initiate an HCAExplorer object.
 x <- HCAExplorer()
 ## View the HCAExplorer object to decide which projects to subset.
 x

 ## Decide to subset first project.
 x <- x[1,]
 x

 ## Get all possible manifest file formats for the project.
 formats <- getManifestFileFormats(x)
 formats

 ## Obtain the manifest for the file using only the first format
 manifest <- getManifest(x, formats[1])
 manifest

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{getManifestFileFormats}} for how to obtain file formats that can
 be used as arguments in this function.
}
