% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getManifestFileFormats}
\alias{getManifestFileFormats}
\alias{getManifestFileFormats,HCAExplorer-method}
\title{Show all possible manifest file formats for current selection}
\usage{
\S4method{getManifestFileFormats}{HCAExplorer}(x)
}
\arguments{
\item{x}{An HCAExplorer object}
}
\value{
A character vecotr of information about possible file formats.
}
\description{
Show all possible manifest file formats for the
 current selection of projects in the HCAExplorer object. To be
 used in conjunction with 'getManifest()'.
}
\examples{
 ## Initiate an HCAExplorer object.
 x <- HCAExplorer()
 ## View the HCAExplorer object to decide which projects to subset.
 x

 ## Decide to subset to just obtain the first two projects.
 x[1:2,]
 x

 ## Get all possible manifest file formats for projects.
 formats <- getManifestFileFormats(x)
 formats

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class
 \code{\link{getManifest}} for how to obtain the manifest file.
}
