% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azul_api.R
\docType{methods}
\name{nextResults}
\alias{nextResults}
\alias{nextResults,HCAExplorer-method}
\title{Next Results}
\usage{
\S4method{nextResults}{HCAExplorer}(result)
}
\value{
A HCAExplorer object that displays the next results
}
\description{
Fetch the next set of bundles from a HCAExplorer Object
}
\examples{

hca <- HCAExplorer()
hca <- nextResults(hca)
hca

}
\author{
Daniel Van Twisk
}
