% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,HCAExplorer-method}
\title{Obtain search results from a HCAExplorer Object}
\usage{
\S4method{results}{HCAExplorer}(object)
}
\value{
a tibble
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated.
}
