% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,HCAExplorer-method}
\title{Obtain search unattenuated results tibble from an HCAExplorer Object}
\usage{
\S4method{results}{HCAExplorer}(x)
}
\arguments{
\item{x}{An HCAExplorer object.}
}
\value{
a tibble giving the unattenuated search results of the HCAExplorer
 object.
}
\description{
Returns a tibble either showing bundles or files based on whichever is
 activated. The tibble returned will display the full results of the search
 instead of the attenuated table shows when dispalying the HCAExplorer
 object.
}
\examples{
 ## Initiate an HCAExplorer object.
 x <- HCAExplorer()
 ## Display the object. Note that the tibble contains very few columns.
 x

 ## Access the tibble and display at available columns using results()
 results(x)

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class.
}
