% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,HCAExplorer-method}
\alias{show,HCAExplorer-method}
\title{Show HCAExplorer}
\usage{
\S4method{show}{HCAExplorer}(object)
}
\arguments{
\item{object}{a HCAExplorer object to show}
}
\value{
outputs a text represntation of the object
}
\description{
Show HCAExplorer
}
\examples{
 ## Initiate an HCAExplorer object.
 x <- HCAExplorer()
 ## Invoke show() by simply displaying the object.
 x

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class.
}
