% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{undoQuery}
\alias{undoQuery}
\alias{undoQuery,HCAExplorer-method}
\title{Undo one or multiple queries performed on an object}
\usage{
\S4method{undoQuery}{HCAExplorer}(x, n = 1L)
}
\arguments{
\item{x}{An HCAExplorer object}

\item{n}{integer(1). The number of queries to step back from.}
}
\value{
An HCAExplorer object with the changes applied to it.
}
\description{
Undo one or multiple queries performed on an HCAExplorer
 object then return the result.
}
\examples{
 ## Initiate an HCAExplorer Object
 x <- HCAExplorer()

 ## First we want to perform a search for certain organs.
 ## Display possible fields looking for organs.
 fields(x)
 ## organs can be queried with "organ".
 ## What values can the field "organ" have?
 values(x, "organ")

 ## Construct a query looking for projects that involve brain.
 x <- x \%>\% filter(organ == brain)
 x

 ## Now select the first two projects by row.
 x <- x[1:2,]
 x

 ## Finally we can remove the previous query using undoQuery().
 ## search.
 x <- undoQuery(x, n = 1)
 x

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{resetQuery}} for reseting the entire query.
}
