% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utilities.R
\name{values,HCAExplorer-method}
\alias{values,HCAExplorer-method}
\title{List all values for certain fields in a HCAExplorer Object}
\usage{
\S4method{values}{HCAExplorer}(x, fields)
}
\arguments{
\item{x}{An HCAExplorer Object.}

\item{fields}{A character vector of fields to display available values for.}
}
\value{
a character vector of possible values for a filter.
}
\description{
To be used in conjunction with the fields function. This
 function allows one to find which values can be queried upon in a given
 field.
}
\examples{
## Intiate an HCAExplorer object.
x <- HCAExplorer()

## Find which firelds can be queried upon.
fields(x)

## Which values can "organ" be searched with?
values(x, 'organ')

}
\seealso{
\code{\link{HCAExplorer}} for the HCAExplorer class,
 \code{\link{fields}} for how how to search for fields in an HCAExplorer object.
}
